requirements.txt
asgiref==3.7.2
boto3==1.34.5
botocore==1.34.5
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-environ==0.11.2
django-storages==1.14.2
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.9.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.0.7
